Do not TSort Finisher and Bootstrap initializers
- Use sorting in topological order (TSort) on all initializers found by Railtie but not on the ones from Rails::Application::Bootstrap and Rails::Application::Finisher. Bootstrap should always be the first and Finisher the last.
- Move Rails::Initializable::Collection to Rails::Application. Now it's only needed there.
